requirements.txt
flask
pycryptodome
python-dotenv
pytest  
pytest-cov 